requirements.txt
# -i https://pypi.tuna.tsinghua.edu.cn/simple
selenium
opencv-python
requests
pyautogui
pytest
allure-pytest
openpyxl
pywinauto